fedotllm/requirements.txt
# Fedot
fedot>=0.7.3.2
thegolem==0.4.0
giotto_tda==0.6.0
ripser==0.6.4

# Data things.
# If you don't have some data, I don't know what you are doing here.
pandas
scipy
tabulate
numpy

# Bugs hunting things
pytest~=8.3.2

# HTTP requests things. You need this even if you use only local llms. Trust me.
requests~=2.32.3
tenacity~=8.5.0

# LangChain.
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-openai==0.2.9
langgraph==0.2.53
langchain-text-splitters==0.3.2
langchain-ollama==0.2.0
langchain-nomic==0.1.4
langchain-chroma==0.1.4

# Web Interface
streamlit~=1.37.1
streamlit-extras~=0.4.7

typing_extensions~=4.12.2
pydantic~=2.8.2
asyncio~=3.4.3
ipython~=8.26.0
graphviz~=0.20.3
scikit-learn~=1.2.2
markdownify
chromadb==0.5.7
nomic[local]
beautifulsoup4~=4.12.3
pyyaml==6.0.2
dynaconf
chromadb
watchdog
loguru
autoflake
langfuse
gradio